feat:issue #625 part 2

+++ service/routing_config_v2.go
@@ -35,14 +35,18 @@ import (
 )
 
 var (
-	// RoutingConfigFilterAttrs router config filter attrs
+	// RoutingConfigV2FilterAttrs router config filter attrs
 	RoutingConfigV2FilterAttrs = map[string]bool{
+		"id":             true,
 		"name":           true,
+		"service":        true,
 		"namespace":      true,
 		"source_service": true,
 		"enable":         true,
 		"offset":         true,
 		"limit":          true,
+		"order_field":    true,
+		"order_type":     true,
 	}
 )
 
@@ -69,13 +73,13 @@ func (s *Server) createRoutingConfigV2(ctx context.Context, req *apiv2.Routing) 
 	// 构造底层数据结构，并且写入store
 	conf, err := api2RoutingConfigV2(req)
 	if err != nil {
-		log.Error("[Service][Routing] parse routing config v2 from request for create",
+		log.Error("[Routing][V2] parse routing config v2 from request for create",
 			utils.ZapRequestIDByCtx(ctx), zap.Error(err))
 		return apiv2.NewResponse(api.ExecuteException)
 	}
 
 	if err := s.storage.CreateRoutingConfigV2(conf); err != nil {
-		log.Error("[Service][Routing] create routing config v2 store layer",
+		log.Error("[Routing][V2] create routing config v2 store layer",
 			utils.ZapRequestIDByCtx(ctx), zap.Error(err))
 		return apiv2.NewResponse(api.StoreLayerException)
 	}
@@ -101,9 +105,12 @@ func (s *Server) DeleteRoutingConfigsV2(ctx context.Context, req []*apiv2.Routin
 
 // DeleteRoutingConfigV2 删除一个路由配置
 func (s *Server) deleteRoutingConfigV2(ctx context.Context, req *apiv2.Routing) *apiv2.Response {
+	if resp := checkRoutingConfigIDV2(req); resp != nil {
+		return resp
+	}
 
 	if err := s.storage.DeleteRoutingConfigV2(req.Id); err != nil {
-		log.Error("[Service][Routing] delete routing config v2 store layer",
+		log.Error("[Routing][V2] delete routing config v2 store layer",
 			utils.ZapRequestIDByCtx(ctx), zap.Error(err))
 		return apiv2.NewResponse(api.StoreLayerException)
 	}
@@ -129,8 +136,11 @@ func (s *Server) UpdateRoutingConfigsV2(ctx context.Context, req []*apiv2.Routin
 
 // updateRoutingConfigV2 更新单个路由配置
 func (s *Server) updateRoutingConfigV2(ctx context.Context, req *apiv2.Routing) *apiv2.Response {
-	// 先判断是不是从 v1 过来的修改操作
 	extendInfo := req.GetExtendInfo()
+	// 如果当前待修改的 v2 路由规则，其实是从 v1 规则在 cache 中转换而来的，则需要先做以下几个步骤
+	// step 1: 将 v1 规则真实的转为 v2 规则
+	// step 2: 将本次要修改的 v2 规则，在 v1 规则中的 inBound 或者 outBound 找到对应的 route，设置其规则 ID
+	// step 3: 进行存储持久化
 	if _, ok := extendInfo[v2.V1RuleIDKey]; ok {
 		resp := s.updateRoutingConfigV2FromV1(ctx, req)
 		if resp.GetCode() != apiv1.ExecuteSuccess {
@@ -138,10 +148,14 @@ func (s *Server) updateRoutingConfigV2(ctx context.Context, req *apiv2.Routing) 
 		}
 	}
 
+	if resp := checkUpdateRoutingConfigV2(req); resp != nil {
+		return resp
+	}
+
 	// 检查路由配置是否存在
 	conf, err := s.storage.GetRoutingConfigV2WithID(req.Id)
 	if err != nil {
-		log.Error("[Service][Routing] get routing config v2 store layer",
+		log.Error("[Routing][V2] get routing config v2 store layer",
 			utils.ZapRequestIDByCtx(ctx), zap.Error(err))
 		return apiv2.NewResponse(api.StoreLayerException)
 	}
@@ -152,13 +166,13 @@ func (s *Server) updateRoutingConfigV2(ctx context.Context, req *apiv2.Routing) 
 	// 作为一个整体进行Update，所有参数都要传递
 	reqModel, err := api2RoutingConfigV2(req)
 	if err != nil {
-		log.Error("[Service][Routing] parse routing config v2 from request for update",
+		log.Error("[Routing][V2] parse routing config v2 from request for update",
 			utils.ZapRequestIDByCtx(ctx), zap.Error(err))
 		return apiv2.NewResponse(api.ExecuteException)
 	}
 
 	if err := s.storage.UpdateRoutingConfigV2(reqModel); err != nil {
-		log.Error("[Service][Routing] update routing config v2 store layer",
+		log.Error("[Routing][V2] update routing config v2 store layer",
 			utils.ZapRequestIDByCtx(ctx), zap.Error(err))
 		return apiv2.NewResponse(apiv1.StoreLayerException)
 	}
@@ -169,36 +183,20 @@ func (s *Server) updateRoutingConfigV2(ctx context.Context, req *apiv2.Routing) 
 
 // GetRoutingConfigsV2 提供给OSS的查询路由配置的接口
 func (s *Server) GetRoutingConfigsV2(ctx context.Context, query map[string]string) *apiv2.BatchQueryResponse {
-	// 先处理offset和limit
-	offset, limit, err := ParseOffsetAndLimit(query)
-	if err != nil {
-		return apiv2.NewBatchQueryResponse(api.InvalidParameter)
+	args, presp := parseRoutingArgs(query, ctx)
+	if presp != nil {
+		return apiv2.NewBatchQueryResponse(presp.GetCode())
 	}
 
-	// 处理剩余的参数
-	filter := make(map[string]string)
-	for key, value := range query {
-		if _, ok := RoutingConfigV2FilterAttrs[key]; !ok {
-			log.Errorf("[Server][Routing][Query] attribute(%s) is not allowed", key)
-			return apiv2.NewBatchQueryResponse(api.InvalidParameter)
-		}
-		filter[key] = value
-	}
-	// service -- > name 这个特殊处理一下
-	if service, ok := filter["service"]; ok {
-		filter["name"] = service
-		delete(filter, "service")
-	}
-
-	total, ret, err := s.Cache().RoutingConfig().GetRoutingConfigsV2(parseRoutingArgs(filter, offset, limit, ctx))
+	total, ret, err := s.Cache().RoutingConfig().GetRoutingConfigsV2(args)
 	if err != nil {
-		log.Error("[Server][Routing] query routing list from cache", utils.ZapRequestIDByCtx(ctx), zap.Error(err))
+		log.Error("[Routing][V2] query routing list from cache", utils.ZapRequestIDByCtx(ctx), zap.Error(err))
 		return apiv2.NewBatchQueryResponse(api.ExecuteException)
 	}
 
 	data, err := marshalRoutingV2toAnySlice(ret)
 	if err != nil {
-		log.Error("[Server][Routing] marshal routing list to anypb.Any list",
+		log.Error("[Routing][V2] marshal routing list to anypb.Any list",
 			utils.ZapRequestIDByCtx(ctx), zap.Error(err))
 		return apiv2.NewBatchQueryResponse(api.ExecuteException)
 	}
@@ -222,10 +220,13 @@ func (s *Server) EnableRoutings(ctx context.Context, req []*apiv2.Routing) *apiv
 }
 
 func (s *Server) enableRoutings(ctx context.Context, req *apiv2.Routing) *apiv2.Response {
+	if resp := checkRoutingConfigIDV2(req); resp != nil {
+		return resp
+	}
 	// 检查路由配置是否存在
 	conf, err := s.storage.GetRoutingConfigV2WithID(req.Id)
 	if err != nil {
-		log.Error("[Service][Routing] get routing config v2 store layer",
+		log.Error("[Routing][V2] get routing config v2 store layer",
 			utils.ZapRequestIDByCtx(ctx), zap.Error(err))
 		return apiv2.NewResponse(api.StoreLayerException)
 	}
@@ -236,7 +237,7 @@ func (s *Server) enableRoutings(ctx context.Context, req *apiv2.Routing) *apiv2.
 	conf.Enable = req.GetEnable()
 
 	if err := s.storage.EnableRouting(conf); err != nil {
-		log.Error("[Service][Routing] enable routing config v2 store layer",
+		log.Error("[Routing][V2] enable routing config v2 store layer",
 			utils.ZapRequestIDByCtx(ctx), zap.Error(err))
 		return apiv2.NewResponse(apiv1.StoreLayerException)
 	}
@@ -246,16 +247,35 @@ func (s *Server) enableRoutings(ctx context.Context, req *apiv2.Routing) *apiv2.
 }
 
 // parseServiceArgs 解析服务的查询条件
-func parseRoutingArgs(filter map[string]string, offset, limit uint32, ctx context.Context) *cache.RoutingArgs {
+func parseRoutingArgs(query map[string]string, ctx context.Context) (*cache.RoutingArgs, *apiv2.Response) {
+	// 先处理offset和limit
+	offset, limit, err := ParseOffsetAndLimit(query)
+	if err != nil {
+		return nil, apiv2.NewResponse(api.InvalidParameter)
+	}
+
+	filter := make(map[string]string)
+	for key, value := range query {
+		if _, ok := RoutingConfigV2FilterAttrs[key]; !ok {
+			log.Errorf("[Routing][V2][Query] attribute(%s) is not allowed", key)
+			return nil, apiv2.NewResponse(api.InvalidParameter)
+		}
+		filter[key] = value
+	}
+
 	res := &cache.RoutingArgs{
-		Filter:    filter,
-		Namespace: filter["namespace"],
-		Offset:    offset,
-		Limit:     limit,
+		Filter:     filter,
+		ID:         filter["id"],
+		Namespace:  filter["namespace"],
+		Service:    filter["service"],
+		OrderField: filter["order_field"],
+		OrderType:  filter["order_type"],
+		Offset:     offset,
+		Limit:      limit,
 	}
 	var ok bool
 	if res.Name, ok = filter["name"]; ok && store.IsWildName(res.Name) {
-		log.Infof("[Service][Routing][Query] fuzzy search with name %s", res.Name)
+		log.Infof("[Routing][V2][Query] fuzzy search with name %s", res.Name)
 		res.FuzzyName = true
 	}
 	if enableStr, ok := filter["enable"]; ok {
@@ -267,7 +287,7 @@ func parseRoutingArgs(filter map[string]string, offset, limit uint32, ctx contex
 		}
 	}
 	log.Infof("[Service][Routing][Query] routing query args: %+v", res)
-	return res
+	return res, nil
 }
 
 // checkBatchRoutingConfig 检查批量请求
@@ -293,13 +313,58 @@ func checkRoutingConfigV2(req *apiv2.Routing) *apiv2.Response {
 		return apiv2.NewRoutingResponse(api.InvalidServiceName, req)
 	}
 
-	// if err := checkResourceName(utils.NewStringValue(req.GetNamespace())); err != nil {
-	// 	return apiv2.NewRoutingResponse(api.InvalidNamespaceName, req)
-	// }
-
 	if err := CheckDbStrFieldLen(utils.NewStringValue(req.GetNamespace()),
 		MaxDbServiceNamespaceLength); err != nil {
 		return apiv2.NewRoutingResponse(api.InvalidNamespaceName, req)
+	}
+
+	if req.GetRoutingPolicy() != apiv2.RoutingPolicy_RulePolicy {
+		return apiv2.NewRoutingResponse(api.InvalidRoutingPolicy, req)
+	}
+
+	// 自动根据 policy 补充 @type 属性
+	if req.RoutingConfig.TypeUrl == "" {
+		if req.GetRoutingPolicy() == apiv2.RoutingPolicy_RulePolicy {
+			req.RoutingConfig.TypeUrl = v2.RuleRoutingTypeUrl
+		}
+		if req.GetRoutingPolicy() == apiv2.RoutingPolicy_MetadataPolicy {
+			req.RoutingConfig.TypeUrl = v2.MetaRoutingTypeUrl
+		}
+	}
+
+	return nil
+}
+
+func checkRoutingConfigIDV2(req *apiv2.Routing) *apiv2.Response {
+	if req == nil {
+		return apiv2.NewRoutingResponse(api.EmptyRequest, req)
+	}
+
+	if req.Id == "" {
+		return apiv2.NewResponse(api.InvalidRoutingID)
+	}
+
+	return nil
+}
+
+// checkUpdateRoutingConfigV2 检查路由配置基础参数有效性
+func checkUpdateRoutingConfigV2(req *apiv2.Routing) *apiv2.Response {
+	if resp := checkRoutingConfigIDV2(req); resp != nil {
+		return resp
+	}
+
+	if req.GetRoutingPolicy() != apiv2.RoutingPolicy_RulePolicy {
+		return apiv2.NewRoutingResponse(api.InvalidRoutingPolicy, req)
+	}
+
+	// 自动根据 policy 补充 @type 属性
+	if req.RoutingConfig.TypeUrl == "" {
+		if req.GetRoutingPolicy() == apiv2.RoutingPolicy_RulePolicy {
+			req.RoutingConfig.TypeUrl = v2.RuleRoutingTypeUrl
+		}
+		if req.GetRoutingPolicy() == apiv2.RoutingPolicy_MetadataPolicy {
+			req.RoutingConfig.TypeUrl = v2.MetaRoutingTypeUrl
+		}
 	}
 
 	return nil
@@ -307,27 +372,24 @@ func checkRoutingConfigV2(req *apiv2.Routing) *apiv2.Response {
 
 // api2RoutingConfig 把API参数转换为内部的数据结构
 func api2RoutingConfigV2(req *apiv2.Routing) (*v2.RoutingConfig, error) {
-
 	out := &v2.RoutingConfig{
-		ID:       req.Id,
-		Revision: req.Revision,
-		Valid:    true,
+		Valid: true,
 	}
 
-	if out.ID == "" {
-		out.ID = utils.NewRoutingV2UUID()
+	if req.Id == "" {
+		req.Id = utils.NewRoutingV2UUID()
 	}
-	if out.Revision == "" {
-		out.Revision = utils.NewV2Revision()
+	if req.Revision == "" {
+		req.Revision = utils.NewV2Revision()
 	}
 
 	if err := out.ParseFromAPI(req); err != nil {
 		return nil, err
 	}
-
 	return out, nil
 }
 
+// marshalRoutingV2toAnySlice 转换为 []*anypb.Any 数组
 func marshalRoutingV2toAnySlice(routings []*v2.ExtendRoutingConfig) ([]*any.Any, error) {
 	ret := make([]*any.Any, 0, len(routings))
